requirements.txt
# Microsoft AI Foundry and Agent Framework
azure-ai-projects>=1.0.0
azure-identity>=1.15.0
azure-ai-inference>=1.0.0b9

# OpenTelemetry for telemetry collection
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# Additional dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0
